+++ src/Components/Molecules/sub-menu.js/sub-menu.js
@@ -1,5 +1,5 @@
 import jsonFile from 'src/db.json';
-import SubMenuBack from 'src/Components/Atoms/buttons/submenu-back-button.js/submenu-back';
+import SubMenuBack from 'src/Components/Atoms/buttons/submenu-back-button/submenu-back';
 import './sub-menu.scss';
 
 const SubMenu = (props) => {
